Silence warnings from the compiler about unused vars and invalid format strings.
The ID event variable in kqh_monitor_process() appears to be unused and should be considered for removal unless there are plans to make use of it in the future.
Calls to rb_raise() which pass strerror(errno) as the format string have been altered to pass strerror(errno) as an argument to "%s" format string. This is slightly more secure as it can prevent overflows should the strerror(errno) contain formating tokens. This style is often recommended in OpenBSD manpages.